Service separated into sub-features for entity reads/writes etc
